requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
axial-positional-embedding==0.3.12
blosc2==3.7.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
coloredlogs==15.0.1
colt5-attention==0.11.1
contourpy==1.3.3
cycler==0.12.1
einops==0.8.1
filelock==3.13.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2024.6.1
gitdb==4.0.12
gitpython==3.1.45
gluonts==0.16.2
hf-xet==1.1.9
huggingface-hub==0.34.4
humanfriendly==10.0
hyper-connections==0.2.1
idna==3.10
igraph==0.11.9
jaxtyping==0.3.2
jinja2==3.1.4
joblib==1.5.2
kiwisolver==1.4.9
lightning-utilities==0.15.2
local-attention==1.11.2
markupsafe==2.1.5
matplotlib==3.10.5
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
ndindex==1.10.0
networkx==3.3
numexpr==2.11.0
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pandas==2.3.2
pillow==11.0.0
platformdirs==4.4.0
product-key-memory==0.2.11
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
ptflops==0.7.5
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic-core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-igraph==0.11.9
pytorch-lightning==2.5.3
pytz==2025.2
pyyaml==6.0.2
reformer-pytorch==1.4.4
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentry-sdk==2.35.1
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tables==3.10.2
texttable==1.7.0
threadpoolctl==3.6.0
toolz==0.12.1
torchdiffeq==0.2.5
torchmetrics==1.8.1
tqdm==4.67.1
triton==3.1.0
typing-extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
wadler-lindig==0.1.7
wandb==0.21.2
yarl==1.20.1